refactor(sdp-api): PRO-1532 standardize ramp providers into per-capability directories with webhook processors
