+++ apps/sdp-api/src/lib/ramps/providers/moonpay/client.ts
@@ -1,6 +1,5 @@
 import type {
   Counterparty,
-  MoonpayRampSettlement,
   PaymentRampEstimate,
   PaymentRampQuote,
   SdpEnvironment,
@@ -11,11 +10,10 @@ import {
   parseFiatCurrency,
 } from "@sdp/types/payment-rails";
 import type { CounterpartyRequirements } from "@sdp/types/ramp-requirements";
-import { AppError, badRequest, providerNotConfigured } from "@/lib/errors";
-import { verifyWebhookSignature } from "@/lib/webhook-signature";
-import { providerFetchJson } from "../fetch";
-import { readyCounterparty } from "../requirements";
-import { createProviderRampSupport, RAMP_RAIL_DUMPS, rampId, requireEnv } from "../shared";
+import { AppError, providerNotConfigured } from "@/lib/errors";
+import { providerFetchJson } from "../../fetch";
+import { readyCounterparty } from "../../requirements";
+import { createProviderRampSupport, RAMP_RAIL_DUMPS, rampId, requireEnv } from "../../shared";
 import type {
   MutableProviderRampSupport,
   ProviderRampSupport,
@@ -26,11 +24,8 @@ import type {
   RampOnrampQuoteInput,
   RampProvider,
   RampRuntimeContext,
-  RampSettlementEvent,
-  RampWebhookValidationContext,
-  RampWebhookValidationResult,
   ValidateCounterpartyOptions,
-} from "../types";
+} from "../../types";
 
 const MOONPAY_API_BASE_URL = "https://api.moonpay.com";
 const MOONPAY_ONRAMP_URL = "https://buy.moonpay.com";
@@ -111,51 +106,6 @@ function normalizeMoonpayCurrencyCode(value: string): string {
     return `${normalized.slice(0, -"_SOLANA".length)}_SOL`.toLowerCase();
   }
   return normalized.toLowerCase();
-}
-
-function readMoonpayWebhookKey(
-  env: Record<string, string | undefined>,
-  environment: SdpEnvironment
-): string {
-  const webhookKey = (
-    environment === "sandbox" ? env.MOONPAY_SANDBOX_WEBHOOK_KEY : env.MOONPAY_WEBHOOK_KEY
-  )?.trim();
-  if (!webhookKey) {
-    throw providerNotConfigured(
-      environment === "sandbox"
-        ? "MoonPay sandbox webhook key is not configured (MOONPAY_SANDBOX_WEBHOOK_KEY)."
-        : "MoonPay webhook key is not configured (MOONPAY_WEBHOOK_KEY)."
-    );
-  }
-  return webhookKey;
-}
-
-function parseMoonpaySignatureV2Header(
-  header: string
-): { timestamp: string; signature: string } | null {
-  let timestamp: string | null = null;
-  let signature: string | null = null;
-
-  for (const part of header.split(",")) {
-    const separatorIndex = part.indexOf("=");
-    if (separatorIndex === -1) {
-      continue;
-    }
-    const prefix = part.slice(0, separatorIndex).trim();
-    const value = part.slice(separatorIndex + 1).trim();
-    if (prefix === "t") {
-      timestamp = value;
-    }
-    if (prefix === "s") {
-      signature = value;
-    }
-  }
-
-  if (!timestamp || !signature) {
-    return null;
-  }
-
-  return { timestamp, signature };
 }
 
 async function moonpaySignature(unsignedQuery: string, secretKey: string): Promise<string> {
@@ -255,82 +205,6 @@ function extractSupport(currencies: readonly MoonpayCurrencyEntry[]): ProviderRa
   return support;
 }
 
-const MOONPAY_TRANSACTION_STATUS = {
-  waitingPayment: "awaiting_payment",
-  pending: "settling",
-  waitingAuthorization: "settling",
-  completed: "settled",
-  failed: "failed",
-} as const satisfies Record<string, RampSettlementEvent["kind"]>;
-type MoonpayTransactionStatus = keyof typeof MOONPAY_TRANSACTION_STATUS;
-
-interface MoonpayTransactionWebhook {
-  type: "transaction_created" | "transaction_updated" | "transaction_failed";
-  data: {
-    id: string;
-    status: MoonpayTransactionStatus;
-    externalTransactionId: string | null;
-    failureReason: string | null;
-    baseCurrencyAmount?: number;
-    quoteCurrencyAmount?: number;
-    feeAmount?: number;
-    extraFeeAmount?: number;
-    networkFeeAmount?: number;
-    areFeesIncluded?: boolean;
-    usdRate?: number;
-    cryptoTransactionId?: string | null;
-    baseCurrency?: { code: string };
-    currency?: { code: string };
-  };
-}
-
-function buildMoonpaySettlement(
-  data: MoonpayTransactionWebhook["data"],
-  status: MoonpayRampSettlement["status"]
-): MoonpayRampSettlement | undefined {
-  const {
-    baseCurrency,
-    currency,
-    baseCurrencyAmount,
-    quoteCurrencyAmount,
-    feeAmount,
-    extraFeeAmount,
-    networkFeeAmount,
-    areFeesIncluded,
-    usdRate,
-    cryptoTransactionId,
-    failureReason,
-  } = data;
-  if (
-    !baseCurrency ||
-    !currency ||
-    baseCurrencyAmount === undefined ||
-    quoteCurrencyAmount === undefined ||
-    feeAmount === undefined ||
-    extraFeeAmount === undefined ||
-    networkFeeAmount === undefined ||
-    areFeesIncluded === undefined ||
-    usdRate === undefined
-  ) {
-    return undefined;
-  }
-  return {
-    provider: "moonpay",
-    status,
-    baseCurrencyCode: baseCurrency.code.toUpperCase(),
-    baseCurrencyAmount,
-    quoteCurrencyCode: currency.code.toUpperCase(),
-    quoteCurrencyAmount,
-    feeAmount,
-    extraFeeAmount,
-    networkFeeAmount,
-    areFeesIncluded,
-    usdRate,
-    ...(cryptoTransactionId ? { cryptoTransactionId } : {}),
-    ...(failureReason ? { failureReason } : {}),
-  };
-}
-
 export class MoonpayRampClient implements RampProvider {
   readonly id = "moonpay";
 
@@ -367,91 +241,6 @@ export class MoonpayRampClient implements RampProvider {
     return extractSupport(
       await readDump<readonly MoonpayCurrencyEntry[]>(RAMP_RAIL_DUMPS.moonpay.currencies.file)
     );
-  }
-
-  async validateWebhook({
-    env,
-    environment,
-    headers,
-    rawBody,
-  }: RampWebhookValidationContext): Promise<RampWebhookValidationResult> {
-    const webhookKey = readMoonpayWebhookKey(env, environment);
-    const signatureHeader = headers.get("moonpay-signature-v2")?.trim();
-    if (!signatureHeader) {
-      throw new AppError("UNAUTHORIZED", "MoonPay webhook is missing Moonpay-Signature-V2 header", {
-        provider: this.id,
-      });
-    }
-
-    const parsed = parseMoonpaySignatureV2Header(signatureHeader);
-    if (!parsed) {
-      throw new AppError("UNAUTHORIZED", "MoonPay webhook signature header is malformed", {
-        provider: this.id,
-      });
-    }
-
-    await verifyWebhookSignature({
-      provider: this.id,
-      signedPayload: `${parsed.timestamp}.${rawBody}`,
-      signature: parsed.signature,
-      algorithm: { type: "hmac-sha256", secret: webhookKey, encoding: "hex" },
-      timestampSeconds: Number(parsed.timestamp),
-    });
-
-    let payload: unknown;
-    try {
-      payload = JSON.parse(rawBody);
-    } catch {
-      throw badRequest("MoonPay webhook body must be valid JSON", {
-        provider: this.id,
-      });
-    }
-
-    return { provider: this.id, payload };
-  }
-
-  parseSettlementEvent(payload: unknown): RampSettlementEvent {
-    const { type, data } = payload as MoonpayTransactionWebhook;
-    if (
-      type !== "transaction_created" &&
-      type !== "transaction_updated" &&
-      type !== "transaction_failed"
-    ) {
-      return { provider: this.id, kind: "ignore", reason: `unsupported_event:${type}` };
-    }
-
-    const reference = data.externalTransactionId;
-    if (!reference) {
-      return { provider: this.id, kind: "ignore", reason: "missing_external_transaction_id" };
-    }
-
-    const kind = MOONPAY_TRANSACTION_STATUS[data.status];
-    if (!kind) {
-      return { provider: this.id, kind: "ignore", reason: `unsupported_status:${data.status}` };
-    }
-    if (kind === "failed") {
-      const settlement = buildMoonpaySettlement(data, "failed");
-      return {
-        provider: this.id,
-        kind,
-        reference,
-        ...(data.failureReason ? { error: data.failureReason } : {}),
-        ...(settlement ? { settlement } : {}),
-      };
-    }
-    if (kind === "settled") {
-      const settlement = buildMoonpaySettlement(data, "completed");
-      return {
-        provider: this.id,
-        kind,
-        reference,
-        ...(data.quoteCurrencyAmount !== undefined
-          ? { receivedAmount: String(data.quoteCurrencyAmount) }
-          : {}),
-        ...(settlement ? { settlement } : {}),
-      };
-    }
-    return { provider: this.id, kind, reference };
   }
 
   async estimateOnramp(
