+++ apps/sdp-api/src/lib/webhook-signature.ts
@@ -5,9 +5,11 @@ const encoder = new TextEncoder();
 
 export const WEBHOOK_TIMESTAMP_TOLERANCE_SECONDS = 300;
 
+type WebhookSignatureEncoding = "base64" | "hex";
+
 type WebhookSignatureAlgorithm =
-  | { type: "hmac-sha256"; secret: string; encoding: "base64" | "hex" }
-  | { type: "ecdsa-sha256"; publicKeyPem: string };
+  | { type: "hmac-sha256"; secret: string; encoding: WebhookSignatureEncoding }
+  | { type: "ecdsa-sha256"; publicKeyPem: string; encoding: WebhookSignatureEncoding };
 
 export interface VerifyWebhookSignatureInput {
   provider: string;
@@ -20,16 +22,14 @@ export interface VerifyWebhookSignatureInput {
   timestampSeconds: number;
 }
 
-function decodeSignature(signature: string, encoding: "base64" | "hex"): Uint8Array | null {
-  if (encoding === "hex") {
-    if (!/^[0-9a-f]+$/i.test(signature) || signature.length % 2 !== 0) {
-      return null;
-    }
-    return Uint8Array.from(Buffer.from(signature, "hex"));
-  }
-  return Uint8Array.from(Buffer.from(signature, "base64"));
-}
-
+/**
+ * Verifies an HMAC-SHA256 webhook signature against the provider's signed payload.
+ *
+ * @param secret - Shared webhook signing secret configured for the provider.
+ * @param signedPayload - Exact string the provider signed, usually the raw body or timestamp-prefixed body.
+ * @param signatureBytes - Decoded signature bytes from the provider header.
+ * @returns Whether the signature matches the payload for the given secret.
+ */
 async function verifyHmacSha256(
   secret: string,
   signedPayload: string,
@@ -45,6 +45,16 @@ async function verifyHmacSha256(
   return crypto.subtle.verify("HMAC", key, signatureBytes, encoder.encode(signedPayload));
 }
 
+/**
+ * Verifies an ECDSA P-256/SHA-256 webhook signature against the provider's signed payload.
+ *
+ * @param publicKeyPem - Provider webhook public key in PEM/SPKI format. Literal "\\n"
+ * sequences are normalized to real newlines before verification because PEM keys are
+ * commonly stored in environment variables that way.
+ * @param signedPayload - Exact string the provider signed, usually the raw request body.
+ * @param signatureBytes - Decoded DER signature bytes from the provider header.
+ * @returns Whether the signature matches the payload for the given public key.
+ */
 function verifyEcdsaSha256(
   publicKeyPem: string,
   signedPayload: string,
@@ -58,6 +68,12 @@ function verifyEcdsaSha256(
     );
 }
 
+/**
+ * Validates a provider webhook timestamp and signature.
+ *
+ * @param input - Provider name, raw signature, signed payload, algorithm details, and signed timestamp.
+ * @returns Resolves when the timestamp is fresh and the signature is valid; throws an unauthorized error otherwise.
+ */
 export async function verifyWebhookSignature(input: VerifyWebhookSignatureInput): Promise<void> {
   const { provider, signature, signedPayload, algorithm, timestampSeconds } = input;
 
@@ -71,16 +87,32 @@ export async function verifyWebhookSignature(input: VerifyWebhookSignatureInput)
     );
   }
 
-  const encoding = algorithm.type === "hmac-sha256" ? algorithm.encoding : "base64";
-  const signatureBytes = decodeSignature(signature, encoding);
-  if (!signatureBytes) {
+  let signatureBytes: Uint8Array;
+  switch (algorithm.encoding) {
+    case "hex":
+      if (!/^[0-9a-f]+$/i.test(signature) || signature.length % 2 !== 0) {
+        throw unauthorized(`Invalid ${provider} webhook signature`);
+      }
+      signatureBytes = Uint8Array.from(Buffer.from(signature, "hex"));
+      break;
+    case "base64":
+      signatureBytes = Uint8Array.from(Buffer.from(signature, "base64"));
+      break;
+  }
+
+  if (signatureBytes.length === 0) {
     throw unauthorized(`Invalid ${provider} webhook signature`);
   }
 
-  const valid =
-    algorithm.type === "hmac-sha256"
-      ? await verifyHmacSha256(algorithm.secret, signedPayload, signatureBytes)
-      : verifyEcdsaSha256(algorithm.publicKeyPem, signedPayload, signatureBytes);
+  let valid: boolean;
+  switch (algorithm.type) {
+    case "hmac-sha256":
+      valid = await verifyHmacSha256(algorithm.secret, signedPayload, signatureBytes);
+      break;
+    case "ecdsa-sha256":
+      valid = verifyEcdsaSha256(algorithm.publicKeyPem, signedPayload, signatureBytes);
+      break;
+  }
 
   if (!valid) {
     throw unauthorized(`Invalid ${provider} webhook signature`);
