+++ apps/sdp-api/src/routes/payments/handlers/ramps/lightspark.ts
@@ -3,6 +3,7 @@ import type { CollectedFieldData } from "@sdp/types/ramp-requirements";
 import type { CounterpartyRow } from "@/db/repositories/counterparty.repository";
 import { badRequest, notFound } from "@/lib/errors";
 import { RAMP_PROVIDER_CLIENTS } from "@/lib/ramps";
+import { buildLightsparkAccountInfo } from "@/lib/ramps/providers/lightspark/counterparty";
 import {
   isLightsparkExternalAccountActive,
   type LightsparkPayoutAccount,
@@ -13,9 +14,8 @@ import {
   readLightsparkData,
   readLightsparkPayoutAccountByKey,
   readLightsparkPayoutAccounts,
-} from "@/lib/ramps/providers/lightspark";
+} from "@/lib/ramps/providers/lightspark/provider-data";
 import type { LightsparkCustomerResolution } from "@/lib/ramps/types";
-import { buildLightsparkAccountInfo } from "@/lib/ramps/validation/lightspark";
 import { getCounterpartiesRepository } from "@/routes/counterparties/context";
 import { type AppContext, rampRuntime } from "../../context";
