+++ packages/sdp-types/src/payment-rails.ts
@@ -1,6 +1,5 @@
+import { RAMP_FIAT_CURRENCIES, type RampFiatCurrency } from "./generated/ramp-support.generated";
 import type { RampProviderId } from "./provider-access";
-
-export type FiatCurrencyCode = string;
 
 export const SOLANA_CRYPTO_RAILS = [
   "sol.solana",
@@ -32,32 +31,23 @@ export function getCryptoRailAssetLabel(assetRail: CryptoRailId): CryptoAssetSym
   return CRYPTO_RAIL_ASSET_LABELS[assetRail];
 }
 
-/** On-chain decimals per crypto asset symbol; fiat falls back to 2 minor units. */
-export const CRYPTO_ASSET_DECIMALS = {
-  SOL: 9,
-  USDC: 6,
-  USDT: 6,
-  USDG: 6,
-  PYUSD: 6,
-} as const satisfies Record<CryptoAssetSymbol, number>;
-
-export interface OnrampPairSupport<FiatCurrency extends string = FiatCurrencyCode> {
+export interface OnrampPairSupport<FiatCurrency extends RampFiatCurrency = RampFiatCurrency> {
   source: FiatCurrency;
   dest: CryptoRailId;
   providers: readonly RampProviderId[];
 }
 
-export interface OfframpPairSupport<FiatCurrency extends string = FiatCurrencyCode> {
+export interface OfframpPairSupport<FiatCurrency extends RampFiatCurrency = RampFiatCurrency> {
   source: CryptoRailId;
   dest: FiatCurrency;
   providers: readonly RampProviderId[];
 }
 
-export function parseFiatCurrency(value: string): FiatCurrencyCode | null {
+export function parseFiatCurrency(value: string): RampFiatCurrency | null {
   const normalized = value.trim().toUpperCase();
-  if (!/^[A-Z]{3}$/.test(normalized)) {
+  if (!RAMP_FIAT_CURRENCIES.includes(normalized as RampFiatCurrency)) {
     return null;
   }
 
-  return normalized;
+  return normalized as RampFiatCurrency;
 }
